Worktree typemap correctness (#58)
* Type-map: fix int SqlDbType, time->TimeOnly, varchar(max) column guard; add matrix+parity harness

* Type-map: support real->float (revive TYPE_FLOAT), split float/real routing

- Tokenizer recognizes `real` as TYPE_FLOAT (distinct from float/double's TYPE_DOUBLE)
- Token.CSharpType/SqlDbType/CSharpValue split TYPE_FLOAT (float/Real/f-suffix) from TYPE_DOUBLE (double/Float)
- Runtime NormalizeType splits real->float, float->double to match build-time key
- Matrix + KeyParity rows for real; new RealTests.cs end-to-end snapshot

* Type-map: add bigint/smallint/tinyint (long/short/byte) end-to-end

* Type-map: add money/smallmoney (decimal, distinct SqlDbType)

* Type-map: add smalldatetime (DateTime) and xml (string)

TYPE_SMALLDATETIME=TYPE+20 mirrors TYPE_DATETIME in every Token.cs arm
(distinct SqlDbType.SmallDateTime); runtime NormalizeType already routed
smalldatetime to System.DateTime. TYPE_XML=TYPE+23 (TYPE+21/22 skipped,
those are TYPE_TABLE/TYPE_OBJECT) mirrors TYPE_STRING for CSharpType/
CSharpValue/reader but emits no length guard in SqlProperty (xml has no
size); added "xml" to runtime NormalizeType's string group. Matrix rows,
KeyParityTests facts, and an end-to-end snapshot added. EditorParityTests
required no changes - all three editor type maps already listed both
types. 249/249 C# tests green, 116/116 VS Code tests green.

* Type-map: add image (byte[], SqlDbType.Image)

* Type-map: shred image input-table columns via hex Convert

Fold TYPE_IMAGE into SQuiLShred.IsBinary so an image column in an input
@Params_ table is captured as nvarchar(max) in the OPENJSON WITH clause and
hex-decoded with Convert(varbinary(max), col, 2) in the SELECT, matching the
binary/varbinary path. Adds a Verify snapshot regression.

* Type-map: add timestamp/rowversion (byte[], output-only) + SP0032 input error across all surfaces

TokenType.TYPE_TIMESTAMP folds into the existing binary arms (DataReader,
SqlDbType, CSharpType, CodeBlock.IsBinary). New SQuiLTimestampInputValidator
flags timestamp/rowversion declared as an input (scalar or table column) as
build error SP0032; mirrored as an editor squiggle in the SSMS extension, the
Visual Studio extension, and the VS Code TS extension. Matrix + parity rows
added for the output path; runtime NormalizeType already mapped
timestamp/rowversion to byte[].

* Add SP0032 input-table-column coverage (self-review follow-up)

Closes a coverage gap found during self-review: the validator and all
three linters already flag timestamp/rowversion table columns, not just
scalars, but only the scalar case had a test. Adds
TimestampInputTableColumnIsError (generator) and the matching TS test.

* Type-map: docs (guide/CLAUDE/SKILL), SP0032 recorded, editor-parity completeness assertion + full verification

- guide.html type table: Time->TimeOnly fix + Real/Money/SmallMoney/
  SmallDateTime/Image/Timestamp/RowVersion rows; re-rendered to all 3 hosts
- CLAUDE.md: record SP0030/SP0031/SP0032 in the diagnostic-ID list, fix both
  stale "Next free: SP0030" lines to SP0033, extend Special Handling with the
  full SQL->C# type map (int/bigint/smallint/tinyint, real vs float, money/
  smallmoney, smalldatetime, xml, image, timestamp/rowversion + SP0032,
  time->TimeOnly)
- SKILL.md (both copies, byte-identical): add new types + SP0032 mistake note
- EditorParityTests: unknown editor-map key is now a hard failure instead of
  a silent continue, so both editor SqlTypeMap.cs files are asserted fully
  covered by the canonical matrix

258/258 C# + 119/119 TS green; VS Code VSIX packaged; SSMS + Visual Studio
VSIXes rebuilt clean.

+++ SQuiL.Tests/BasicIODeclareTests/FullVariable/BasicIODeclareTests.FullVariable#FullVariableDataContext.g.verified.cs
@@ -29,7 +29,7 @@ partial class FullVariableDataContext : SQuiLBaseDataContext
 		List<DbParameter> parameters = new()
 		{
 			CreateParameter("@Debug", System.Data.SqlDbType.Bit, request.Debug || EnvironmentName != "Production"),
-			CreateParameter("@Param_Scaler", System.Data.SqlDbType.BigInt, request.Scaler)
+			CreateParameter("@Param_Scaler", System.Data.SqlDbType.Int, request.Scaler)
 		};
 		
 		command.CommandText = Query(parameters);
