+++ SQuiL.Tests/BasicIODeclareTests/TranscriptProcessing/BasicIODeclareTests.TranscriptProcessing#TranscriptProcessingDataContext.g.verified.cs
@@ -30,7 +30,7 @@ partial class TranscriptProcessingDataContext : SQuiLBaseDataContext
 		{
 			CreateParameter("@Debug", System.Data.SqlDbType.Bit, request.Debug || EnvironmentName != "Production"),
 			CreateParameter("@Param_IncludeCompletedTranscripts", System.Data.SqlDbType.Bit, request.IncludeCompletedTranscripts),
-			CreateParameter("@Param_ReferenceID", System.Data.SqlDbType.BigInt, request.ReferenceID),
+			CreateParameter("@Param_ReferenceID", System.Data.SqlDbType.Int, request.ReferenceID),
 			CreateParameter("@Param_IssueDate", System.Data.SqlDbType.DateTime, request.IssueDate)
 		};
